requirements.txt
streamlit==1.51.0
pandas==2.3.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.41
requests==2.32.3
openpyxl==3.1.5